Organiza melhor tipos e arquivos
