+++ .gitignore
@@ -1,1 +1,5 @@
 .DS_Store
+data_processing/neo4j/data
+vizcolab
+vizcolab_db_backup
+*.csv
